Practical_01
Rename Develop a Database “Students” of an Institute. Create a table with following attributes: a)	Roll Number b)	Name of the student c)	Department d)	Year e)	Section f)	Contact Number g)	eMail ID Add another other attributes as “CGPA” in the above mentioned table. Insert details of at-least 10 students taken from the end user in the table and display the detailed table. Generate the sorted reports based on the following criteria: a)	List of students with roll numbers in increasing order b)	List of the students with decreasing order of CGPA. c)	List the students with the Minimum and maximum CGPA. to Practical_01