Practical_04
Examine the provided image “moonlanding.png”, which is heavily contaminated with periodic noise. Clean up the noise using the Fast Fourier Transform (FFT). Find and use the 2-D FFT function in “scipy.fftpack”, and plot the FFT spectrum of the image. The spectrum consists of high and low frequency components. The noise is contained in the high-frequency part of the spectrum, so set some of those components to zero (Hint: use array slicing). Apply the inverse Fourier transform to see the resulting image.