Practical_06
Import sales data of a Company. Analyze it following ways with visualization using 
          Matplotlib:
a)	Read the total profit of all the months and visualize it using a Line Plot.
b)	Read all product sales data and show it using a Multiline Plot.
c)	Read face cream and facewash product sales data and show it using the Bar chart.
Calculate total sale data for last year for each product and show it using a Pie chart.